Move config mgmt from background.ts to own file
